V1 Crud with Users

+++ Services/EFCoreTodoService.cs
@@ -10,30 +10,31 @@ namespace TodoListApi.Services
         private TodoOutputModel ToOutputModel(Todo todo)
            =>  new(todo.Id, todo.Title, todo.CreatedDate, todo.IsActive);
 
-        public async Task<List<TodoOutputModel>> GetAll()
+        public async Task<List<TodoOutputModel>> GetAll(int userId)
         {
-            var todos = await _context.Todos.ToListAsync();
+            var todos = await _context.Todos.Where(t => t.UserId == userId).ToListAsync();
             var outputTodos = todos.ConvertAll(ToOutputModel);
             return outputTodos;
             
         }
-        public async Task<TodoOutputModel?> GetById(int id)
+        public async Task<TodoOutputModel?> GetById(int id, int userId)
         {
-            var todo = await _context.Todos.Where(t =>  t.Id == id).FirstOrDefaultAsync();
+            var todo = await _context.Todos.FirstOrDefaultAsync(t => t.Id == id && t.UserId == userId);
             if (todo is null) return null;
             return ToOutputModel(todo);
         }
-        public async Task<List<TodoOutputModel>> GetActives()
+        public async Task<List<TodoOutputModel>> GetActives(int userId)
         {
-            var todos = await _context.Todos.Where(t => t.IsActive == true).ToListAsync();
+            var todos = await _context.Todos.Where(t => t.IsActive == true && t.UserId == userId).ToListAsync();
             var outputTodos = todos.ConvertAll(ToOutputModel);
             return outputTodos;
         }
 
-        public async Task<TodoOutputModel> Create(TodoInputModel newTodo)
+        public async Task<TodoOutputModel> Add(TodoInputModel newTodo, int userID)
         {
             var dbTodo = new Todo
             {
+                UserId = userID,
                 Title = newTodo.Title,
                 CreatedDate = newTodo.CreatedDate.GetValueOrDefault(),
                 IsActive = newTodo.IsActive.GetValueOrDefault()
@@ -43,18 +44,22 @@ namespace TodoListApi.Services
             return ToOutputModel(dbTodo);
         }
 
-        public async Task<bool> Delete(int id)
+        public async Task<bool> Delete(int id, int userId)
         {
-            return await _context.Todos.Where(t => t.Id == id).ExecuteDeleteAsync() > 0;
+            return await _context.Todos.Where(t => t.Id == id && t.UserId == userId).ExecuteDeleteAsync() > 0;
         }
 
-        public async Task<bool> Update(int id, TodoInputModel newTodo)
+        public async Task<bool> Update(int id, int userId, TodoInputModel newTodo)
         {
-            return await _context.Todos.Where(t => t.Id == id)
-                .ExecuteUpdateAsync(tod => tod
-                    .SetProperty(todo => todo.Title, newTodo.Title)
-                    .SetProperty(todo => todo.CreatedDate, newTodo.CreatedDate)
-                    .SetProperty(todo => todo.IsActive, newTodo.IsActive)) >0;
+            var dbTodo = await _context.Todos.FirstOrDefaultAsync(t => t.Id == id && t.UserId == userId);
+            if (dbTodo is null) return false;
+
+            dbTodo.Title = newTodo.Title;
+            dbTodo.CreatedDate = newTodo.CreatedDate.GetValueOrDefault();
+            dbTodo.IsActive = newTodo.IsActive.GetValueOrDefault();
+
+            _context.Todos.Update(dbTodo);
+            return await _context.SaveChangesAsync() > 0;
         }
     }
 }
